Create new component and add metadata
